feat: replace types to size_t, fix tests and moving size logic to multitypeQueue

+++ src/agent/multitype_queue/include/multitype_queue.hpp
@@ -137,5 +137,5 @@ public:
     int storedItems(MessageType type, const std::string moduleName = "") override;
 
     //TODO: doc
-    int sizePerType(MessageType type) override;
+    size_t sizePerType(MessageType type) override;
 };
